Added new ProgressBar interface
